Enable building Tcl with Microsoft's latest compiler offering (VS2005).  We have to handle a number of oddities as they have deprecated most of the standard C library and now generate manifest files to be linked into the binaries. SF bug #1424909

+++ generic/tclDate.c
@@ -10,7 +10,7 @@
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tclDate.c,v 1.20.4.2 2005/11/04 20:15:09 kennykb Exp $
+ * RCS: @(#) $Id: tclDate.c,v 1.20.4.3 2006/06/14 15:21:14 patthoyts Exp $
  */
 
 #include "tclInt.h"
@@ -478,7 +478,7 @@ Convert(Month, Day, Year, Hours, Minutes, Seconds, Meridian, DSTmode, TimePtr)
         for (i = EPOCH; i < Year; i++)
             Julian += 365 + IsLeapYear(i);
     } else {
-        for (i = Year; i < EPOCH; i++)
+        for (i = (int)Year; i < EPOCH; i++)
             Julian -= 365 + IsLeapYear(i);
     }
     Julian *= SECSPERDAY;
@@ -547,7 +547,7 @@ NamedMonth(Start, MonthOrdinal, MonthNumber)
      *  doing next february from january gives us february of the current year)
      * set day to 1, time to 0
      */
-    tm->tm_year += MonthOrdinal;
+    tm->tm_year += (int)MonthOrdinal;
     if (tm->tm_mon < MonthNumber - 1) {
 	tm->tm_year--;
     }
